agents1014/requirements.txt
# Core dependencies for TRS Multi-Agent Chess System
# 100% local, sovereign architecture

# Chess engine
python-chess==1.999

# Local LLM orchestration
ollama==0.1.7
langchain==0.1.0
langchain-community==0.0.10

# WebSocket communication
websockets==12.0
aiohttp==3.9.1

# Voice interface - local only
openai-whisper==20231117
piper-tts==1.2.0
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1

# Math and transformations
numpy==1.26.2
scipy==1.11.4
sympy==1.12

# Data handling
pydantic==2.5.3
pyyaml==6.0.1

# Async support
asyncio==3.4.3

# Logging and monitoring
structlog==24.1.0

# Configuration
python-dotenv==1.0.0
